+++ src/main/java/chess/repository/dao/entity/GameEntity.java
@@ -1,4 +1,4 @@
-package chess.entity;
+package chess.repository.dao.entity;
 
 import chess.model.game.ChessGame;
