+++ src/test/java/chess/dao/springjdbc/SpringGameDaoTest.java
@@ -3,7 +3,8 @@ package chess.dao.springjdbc;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.entity.GameEntity;
+import chess.repository.dao.entity.GameEntity;
+import chess.repository.dao.springjdbc.SpringGameDao;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
